renaming meta cache file
