+++ renderer/tools/cache-meta.js
